// src/02-http/interfaces.ts
// 01-requests
export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

// 02-refactored
export interface Post {
  userId: number;
  id: number;
  title: string;
  body: string;
}